+++ src/actions/search.js
@@ -1,0 +1,40 @@
+const ACTIONS = {
+    UPDATE_DIFFICULTY_FILTER: "UPDATE_DIFFICULTY_FILTER",
+    ADD_CATEGORY_FILTER: "ADD_CATEGORY_FILTER",
+    REMOVE_CATEGORY_FILTER: "REMOVE_CATEGORY_FILTER",
+    CHANGE_CATEGORY_FILTER_TYPE: "CHANGE_CATEGORY_FILTER_TYPE"
+};
+
+export const SEARCH_ACTIONS = ACTIONS;
+
+export const DIFFICULTY = {
+    EASY: "easy",
+    MEDIUM: "medium",
+    HARD: "hard",
+    ALL: "all"
+};
+
+export const CATEGORY_FILTER_TYPE = {
+    AND: "and",
+    OR: "or"
+};
+
+export const updateDifficultyFilter = difficulty => ({
+    type: ACTIONS.UPDATE_DIFFICULTY_FILTER,
+    difficulty: difficulty
+});
+
+export const addCategoryFilter = category => ({
+    type: ACTIONS.ADD_CATEGORY_FILTER,
+    category: category
+});
+
+export const removeCategoryFilter = category => ({
+    type: ACTIONS.REMOVE_CATEGORY_FILTER,
+    category: category
+});
+
+export const changeCategoryFilterType = filter_type => ({
+    type: ACTIONS.CHANGE_CATEGORY_FILTER_TYPE,
+    filter_type: filter_type
+});
