+++ src/sagas/exercises.js
@@ -1,20 +1,17 @@
 import {put, take, takeEvery} from 'redux-saga/effects';
+import {ACTIONS} from '../actions';
+import {checkWeb3Account} from '../actions/web3';
 import {
-    ACTIONS,
-    checkWeb3Account,
     compile,
-    deploy, EXERCISE_STATE,
+    deploy,
+    EXERCISE_STATE,
     loadCompiler,
     setExerciseError,
     setExerciseUpdate,
     testContracts
-} from '../actions';
+} from '../actions/exercise';
 
-const exercises = [
-    takeEvery(ACTIONS.RUN_EXERCISE, workerExecuteExercise)
-];
-
-export default exercises;
+export default [takeEvery(ACTIONS.RUN_EXERCISE, workerExecuteExercise)];
 
 const exerciseDeployedAction = (actionCodeId, target) =>
     target.codeId === actionCodeId && target.type === ACTIONS.DEPLOY_CONTRACTS_SUCCESS;
@@ -36,7 +33,6 @@ function* workerExecuteExercise(action) {
         yield put(setExerciseUpdate(action.codeId, "Loading compiler"));
         yield put(loadCompiler(action.compilerVersion));
         const compilerAction = yield take(target => compilerLoaded(action.compilerVersion, target));
-        console.log('Compiler loaded.');
 
         yield put(setExerciseUpdate(action.codeId, "Checking wallet access", EXERCISE_STATE.AUTHORIZING));
         yield put(checkWeb3Account());
@@ -57,7 +53,6 @@ function* workerExecuteExercise(action) {
         if (compilationResultAction.type === ACTIONS.COMPILE_FAILURE) {
             return console.log("Compilation failed. Cancel exercise now.");
         }
-        console.log('Exercise compiled');
 
         yield put(deploy(action.codeId, compilationResultAction.code.contracts));
         const deploymentResultAction = yield take([
@@ -67,7 +62,6 @@ function* workerExecuteExercise(action) {
         if (deploymentResultAction.type === ACTIONS.DEPLOY_CONTRACTS_FAILURE) {
             return console.log("Deployment failed. Cancel exercise now.");
         }
-        console.log("Exercise deployed");
 
         yield put(testContracts(action.codeId, action.validation, deploymentResultAction.addresses));
         const testResultAction = yield take([
@@ -77,7 +71,6 @@ function* workerExecuteExercise(action) {
         if (testResultAction.type === ACTIONS.TEST_CONTRACT_FAILURE) {
             return console.log("Tests of exercise failed.");
         }
-        console.log("Exercise completed");
 
     } catch (error) {
         console.log('Error in workerExecuteExercise', action.codeId, error);
